Refactoring + necessary changes in POM to make it build. 	modified:   api/pom.xml 	renamed:    api/src/main/java/org/openmrs/module/mksreports/MekomSolutionsReports.java -> api/src/main/java/org/openmrs/module/mksreports/MKSReports.java 	renamed:    api/src/main/java/org/openmrs/module/mksreports/MekomSolutionsReportsActivator.java -> api/src/main/java/org/openmrs/module/mksreports/MKSReportsActivator.java 	renamed:    api/src/main/java/org/openmrs/module/mksreports/api/MekomSolutionsReportsService.java -> api/src/main/java/org/openmrs/module/mksreports/api/MKSReportsService.java 	renamed:    api/src/main/java/org/openmrs/module/mksreports/api/db/MekomSolutionsReportsDAO.java -> api/src/main/java/org/openmrs/module/mksreports/api/db/MKSReportsDAO.java 	renamed:    api/src/main/java/org/openmrs/module/mksreports/api/db/hibernate/HibernateMekomSolutionsReportsDAO.java -> api/src/main/java/org/openmrs/module/mksreports/api/db/hibernate/HibernateMKSReportsDAO.java 	renamed:    api/src/main/java/org/openmrs/module/mksreports/api/impl/MekomSolutionsReportsServiceImpl.java -> api/src/main/java/org/openmrs/module/mksreports/api/impl/MKSReportsServiceImpl.java 	renamed:    api/src/main/resources/MekomSolutionsReports.hbm.xml -> api/src/main/resources/MKSReports.hbm.xml 	modified:   api/src/main/resources/moduleApplicationContext.xml 	renamed:    api/src/test/java/org/openmrs/module/mksreports/api/MekomSolutionsReportsServiceTest.java -> api/src/test/java/org/openmrs/module/mksreports/api/MKSReportsServiceTest.java 	new file:   omod/.DS_Store 	modified:   omod/src/main/resources/config.xml 	modified:   pom.xml
