+++ api/src/main/java/org/openmrs/module/mksreports/MKSReports.java
@@ -20,7 +20,7 @@ import org.openmrs.BaseOpenmrsMetadata;
 /**
  * It is a model class. It should extend either {@link BaseOpenmrsObject} or {@link BaseOpenmrsMetadata}.
  */
-public class MekomSolutionsReports extends BaseOpenmrsObject implements Serializable {
+public class MKSReports extends BaseOpenmrsObject implements Serializable {
 
 	private static final long serialVersionUID = 1L;
