+++ api/src/main/resources/moduleApplicationContext.xml
@@ -13,4 +13,28 @@
 
 	<context:component-scan base-package="org.openmrs.module.mksreports" />
 
+	<bean id="mksReportsService" class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
+		<property name="transactionManager">
+			<ref bean="transactionManager" />
+		</property>
+		<property name="target">
+			<bean class="org.openmrs.module.mksreports.api.impl.MKSReportsServiceImpl" />
+		</property>
+		<property name="preInterceptors">
+			<ref bean="serviceInterceptors" />
+		</property>
+		<property name="transactionAttributeSource">
+			<ref bean="transactionAttributeSource" />
+		</property>
+	</bean>
+
+	<bean parent="serviceContext">
+		<property name="moduleService">
+			<list merge="true">
+				<value>org.openmrs.module.mksreports.api.MKSReportsService</value>
+				<ref bean="mksReportsService"/>
+			</list>
+		</property>
+	</bean>
+
 </beans>
